Update christmasTree.js
limited the positions ornaments can be placed, not super precise yet. also added in a go back button that is draw on the tree screen to bring you back to the splash screen, functionality is currently limited to changing the scene to splash but stops working there and doesn't actually draw the screen.